New Final Project/requirements.txt
absl-py==2.3.0
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
arcade==3.1.0
asyncua==1.1.6
attrs==25.3.0
av==12.3.0
backoff==2.2.1
boto3==1.35.60
botocore==1.35.99
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
datasets==3.6.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt-ng==0.9.0
einops==0.8.1
ema-pytorch==0.7.7
Farama-Notifications==0.0.4
fastapi==0.110.3
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
greenlet==3.2.2
grpcio==1.71.0
gymnasium==0.29.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
huggingface-sb3==3.0
humanfriendly==10.0
idna==3.7
ifaddr==0.2.0
imageio==2.37.0
imitation==1.0.1
inference==0.50.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpickle==4.1.0
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
munch==4.0.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-ml-py==12.575.51
onnxruntime==1.21.1
openai==1.82.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optuna==4.3.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.0.0
pillow_heif==0.22.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.22.0
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
PyAutoGUI==0.9.54
pybase64==1.0.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
pygame==2.3.0
PyGetWindow==0.0.9
pyglet==2.1.5
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pylogix==1.0.5
pymodbus==3.8.3
PyMsgBox==1.0.9
pymunk==6.9.0
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytiled_parser==2.2.9
pytorch-fid==0.3.0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.8
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.64
rubicon-objc==0.5.0
s3transfer==0.10.4
sacred==0.8.7
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seals==0.2.1
shapely==2.0.7
shellingham==1.5.4
shortuuid==1.0.11
six==1.17.0
slack_sdk==3.33.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
stable-baselines3==2.2.1
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.26
tldextract==5.1.3
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
twilio==9.3.8
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wait_for2==0.3.2
wasabi==1.1.3
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zxing-cpp==2.2.0
